Solution zonder Crane...

+++ src/main/java/CraneMovement.java
@@ -13,7 +13,7 @@ public class CraneMovement {
         this.crane = crane;
     }
 
-    public int travelTime() {
+    public double travelTime() {
         double x_time_component = start.getXdistance(end)/crane.getVx();
         double y_time_component = start.getYdistance(end)/crane.getVy();
         return Math.max((int)y_time_component, (int) x_time_component);
